feat(consultant): rebuild simple pages on enterprise primitives
- earnings: EarningStatus from @prisma/client + session-labels badges
  (fixes PENDING_TRUST rendering undefined at runtime); raw table ->
  ResponsiveTable; 'Org Trust' filter tab + pendingTrustEarnings summary
  bucket (additive earnings-service field, excluded from payable totals)
- referrals: loading/error/empty states for all three queries with
  refetch retry (kills perpetual 'Generating link…'); Qualified stat now
  counts QUALIFIED + REWARDED; all 5 ReferralStatus values badged; dead
  commented 'Program Details' block deleted; ResponsiveTable
- recordings: manual useState+fetch -> react-query (fetcher + query key
  in dashboard-queries, keepPreviousData); recordingStatusBadge
- chats: DashboardHeader scaffold + new shared ChatUnavailable fallback
  (no more blank box on Stream init failure); hardcoded calc() height
  removed
- documents: page-header -> DashboardHeader; documentReviewStatusBadge
- trials: local statusColors map -> trialStatusBadge
- collaborations: InvitationsPanel (1284 lines) decomposed into
  PendingInvitationCard / ActiveCollaborationCard / HostedPlanCard /
  ScheduleSummaries + shared types/format modules; dialogs ->
  ResponsiveModal; external component API unchanged
- home: token sweep, shared resolveSponsoringOrgName, EmptyState+refetch
  error handling
- gray-* -> zinc/semantic-token sweep across all nine pages

Part of the consultant/consultee dashboard redesign (PR-3 of 9).
