+++ app/dashboard/consultant/[consultantId]/(features)/chats/page.tsx
@@ -2,8 +2,11 @@
 
 import { use } from "react";
 import { useQuery } from "@tanstack/react-query";
+import { AlertCircle } from "lucide-react";
 import { DashboardErrorBoundary } from "@/components/DashboardErrorBoundary";
 import { ChatSkeleton } from "@/components/dashboard/DashboardSkeletons";
+import { EmptyState } from "@/components/dashboard/DataCard";
+import { Button } from "@/components/ui/button";
 import { createConsultantQueries } from "@/lib/dashboard-queries";
 import { ChatsTab } from "./ChatsTab";
 
@@ -16,7 +19,12 @@ export default function ChatsPage({
 
   // Use the centralized query configuration
   const detailsQuery = createConsultantQueries(consultantId).details;
-  const { data: consultantDetails, isLoading, error } = useQuery(detailsQuery);
+  const {
+    data: consultantDetails,
+    isLoading,
+    error,
+    refetch,
+  } = useQuery(detailsQuery);
 
   if (isLoading) {
     return <ChatSkeleton />;
@@ -25,21 +33,19 @@ export default function ChatsPage({
   if (error) {
     return (
       <DashboardErrorBoundary>
-        <div className="flex items-center justify-center min-h-[400px]">
-          <div className="p-4 bg-red-50 text-red-600 rounded-lg max-w-md text-center">
-            <h3 className="font-semibold mb-2">Error Loading Chat Data</h3>
-            <p className="text-sm">
-              {error.message ||
-                "Failed to load consultant details for chat. Please try again."}
-            </p>
-            <button
-              onClick={() => window.location.reload()}
-              className="mt-4 px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700"
-            >
+        <EmptyState
+          icon={AlertCircle}
+          title="Error loading chat data"
+          description={
+            error.message ||
+            "Failed to load consultant details for chat. Please try again."
+          }
+          action={
+            <Button variant="outline" onClick={() => refetch()}>
               Retry
-            </button>
-          </div>
-        </div>
+            </Button>
+          }
+        />
       </DashboardErrorBoundary>
     );
   }
@@ -47,14 +53,11 @@ export default function ChatsPage({
   if (!consultantDetails?.user) {
     return (
       <DashboardErrorBoundary>
-        <div className="flex items-center justify-center min-h-[400px]">
-          <div className="p-4 bg-orange-50 text-orange-600 rounded-lg max-w-md text-center">
-            <h3 className="font-semibold mb-2">User Data Not Found</h3>
-            <p className="text-sm">
-              User details not found for the given consultant ID.
-            </p>
-          </div>
-        </div>
+        <EmptyState
+          icon={AlertCircle}
+          title="User data not found"
+          description="User details not found for the given consultant ID."
+        />
       </DashboardErrorBoundary>
     );
   }
