+++ app/dashboard/consultant/[consultantId]/(features)/recordings/components/RecordingCard.tsx
@@ -7,8 +7,6 @@ import {
   Play,
   Clock,
   Calendar,
-  HardDrive,
-  Cloud,
   AlertCircle,
   Loader2,
   Download,
@@ -17,7 +15,8 @@ import {
 } from "lucide-react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
+import { StatusBadge } from "@/components/dashboard/StatusBadge";
+import { recordingStatusBadge } from "@/lib/labels/session-labels";
 import {
   Tooltip,
   TooltipContent,
@@ -74,48 +73,6 @@ export function RecordingCard({ recording, onTransfer }: RecordingCardProps) {
     return `${mins}m`;
   };
 
-  const getStatusBadge = () => {
-    switch (recording.status) {
-      case "AVAILABLE":
-        return (
-          <Badge className="bg-green-500/10 text-green-600 border-green-500/20">
-            <Cloud className="w-3 h-3 mr-1" />
-            Permanent
-          </Badge>
-        );
-      case "READY":
-        return (
-          <Badge className="bg-blue-500/10 text-blue-600 border-blue-500/20">
-            <HardDrive className="w-3 h-3 mr-1" />
-            Stream Storage
-          </Badge>
-        );
-      case "TRANSFERRING":
-        return (
-          <Badge className="bg-yellow-500/10 text-yellow-600 border-yellow-500/20">
-            <Loader2 className="w-3 h-3 mr-1 animate-spin" />
-            Transferring
-          </Badge>
-        );
-      case "FAILED":
-        return (
-          <Badge variant="destructive">
-            <AlertCircle className="w-3 h-3 mr-1" />
-            Failed
-          </Badge>
-        );
-      case "EXPIRED":
-        return (
-          <Badge variant="secondary">
-            <AlertCircle className="w-3 h-3 mr-1" />
-            Expired
-          </Badge>
-        );
-      default:
-        return <Badge variant="outline">{recording.status}</Badge>;
-    }
-  };
-
   const handleTransfer = async () => {
     if (!onTransfer) return;
 
@@ -162,7 +119,7 @@ export function RecordingCard({ recording, onTransfer }: RecordingCardProps) {
               </p>
             )}
           </div>
-          {getStatusBadge()}
+          <StatusBadge {...recordingStatusBadge(recording.status)} />
         </div>
       </CardHeader>
       <CardContent className="space-y-4">
