+++ app/dashboard/consultant/[consultantId]/(features)/referrals/page.tsx
@@ -7,16 +7,28 @@ import {
   DashboardContent,
   DashboardGrid,
 } from "@/components/dashboard/DashboardShell";
-import { StatCard } from "@/components/dashboard/StatCard";
-import { Users, Gift, IndianRupee, Copy, Check, Mail } from "lucide-react";
+import { StatCard, StatCardSkeleton } from "@/components/dashboard/StatCard";
+import { EmptyState, DataCardSkeleton } from "@/components/dashboard/DataCard";
+import { StatusBadge } from "@/components/dashboard/StatusBadge";
+import { referralStatusBadge } from "@/lib/labels/session-labels";
+import {
+  ResponsiveTable,
+  type ResponsiveColumn,
+} from "@/components/ui/responsive-table";
+import {
+  Users,
+  Gift,
+  IndianRupee,
+  Copy,
+  Check,
+  Mail,
+  AlertTriangle,
+} from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { useToast } from "@/hooks/use-toast";
 import { WhatsAppIcon } from "@/components/icons/WhatsAppIcon";
-import {
-  QUALIFICATION_WINDOW_DAYS,
-  CREDIT_EXPIRY_DAYS,
-} from "@/lib/referrals/constants";
+import { QUALIFICATION_WINDOW_DAYS } from "@/lib/referrals/constants";
 
 interface ReferralCode {
   id: string;
@@ -53,6 +65,27 @@ interface CreditData {
   }[];
 }
 
+const creditSourceLabels: Record<string, string> = {
+  REFERRAL_BONUS: "Referral Bonus",
+  REFEREE_BONUS: "Referee Bonus",
+  MANUAL_ADJUSTMENT: "Manual Adjustment",
+};
+
+function formatDate(dateStr: string): string {
+  return new Date(dateStr).toLocaleDateString("en-IN", {
+    day: "numeric",
+    month: "short",
+    year: "numeric",
+  });
+}
+
+function formatAmount(paise: number) {
+  return new Intl.NumberFormat("en-IN", {
+    style: "currency",
+    currency: "INR",
+  }).format(paise / 100);
+}
+
 export default function ConsultantReferralsPage({
   params,
 }: {
@@ -64,7 +97,12 @@ export default function ConsultantReferralsPage({
   const [copied, setCopied] = useState(false);
   const [customCode, setCustomCode] = useState("");
 
-  const { data: codeData } = useQuery<{ data: ReferralCode }>({
+  const {
+    data: codeData,
+    isLoading: codeLoading,
+    error: codeError,
+    refetch: refetchCode,
+  } = useQuery<{ data: ReferralCode }>({
     queryKey: ["referral-code"],
     queryFn: async () => {
       const res = await fetch("/api/referrals/code", { method: "POST" });
@@ -74,7 +112,12 @@ export default function ConsultantReferralsPage({
     staleTime: 60_000,
   });
 
-  const { data: referralsData } = useQuery<{ data: Referral[] }>({
+  const {
+    data: referralsData,
+    isLoading: referralsLoading,
+    error: referralsError,
+    refetch: refetchReferrals,
+  } = useQuery<{ data: Referral[] }>({
     queryKey: ["referrals"],
     queryFn: async () => {
       const res = await fetch("/api/referrals");
@@ -84,7 +127,12 @@ export default function ConsultantReferralsPage({
     staleTime: 30_000,
   });
 
-  const { data: creditsData } = useQuery<{ data: CreditData }>({
+  const {
+    data: creditsData,
+    isLoading: creditsLoading,
+    error: creditsError,
+    refetch: refetchCredits,
+  } = useQuery<{ data: CreditData }>({
     queryKey: ["referral-credits"],
     queryFn: async () => {
       const res = await fetch("/api/referrals/credits");
@@ -98,7 +146,11 @@ export default function ConsultantReferralsPage({
   const referrals = referralsData?.data ?? [];
   const credits = creditsData?.data;
   const totalReferred = referrals.length;
-  const qualified = referrals.filter((r) => r.status === "REWARDED").length;
+  // A referral counts as qualified once it completes the paid booking
+  // (QUALIFIED) — REWARDED is the same milestone after the credit lands.
+  const qualified = referrals.filter(
+    (r) => r.status === "QUALIFIED" || r.status === "REWARDED",
+  ).length;
   const referralLink = code
     ? `${window.location.origin}/r/${code.customCode || code.code}`
     : "";
@@ -136,41 +188,68 @@ export default function ConsultantReferralsPage({
     }
   };
 
-  const referralStatusConfig: Record<
-    string,
-    { label: string; className: string }
-  > = {
-    SIGNED_UP: {
-      label: "Signed Up",
-      className: "bg-yellow-100 text-yellow-700",
+  const statsLoading = referralsLoading || creditsLoading || codeLoading;
+  const statsError = referralsError || creditsError;
+
+  const referralColumns: ResponsiveColumn<Referral>[] = [
+    {
+      key: "user",
+      header: "User",
+      primary: true,
+      cell: (ref) => ref.referredUser.name,
     },
-    REWARDED: {
-      label: "Qualified",
-      className: "bg-green-100 text-green-700",
+    {
+      key: "status",
+      header: "Status",
+      cell: (ref) => <StatusBadge {...referralStatusBadge(ref.status)} />,
     },
-    EXPIRED: { label: "Expired", className: "bg-zinc-100 text-zinc-600" },
-  };
+    {
+      key: "signedUp",
+      header: "Signed Up",
+      cell: (ref) => (
+        <span className="text-zinc-500">{formatDate(ref.signedUpAt)}</span>
+      ),
+    },
+    {
+      key: "reward",
+      header: "Reward",
+      headClassName: "text-right",
+      className: "text-right",
+      cell: (ref) =>
+        ref.status === "REWARDED"
+          ? formatAmount(ref.referrerRewardAmount)
+          : "-",
+    },
+  ];
 
-  const creditSourceLabels: Record<string, string> = {
-    REFERRAL_BONUS: "Referral Bonus",
-    REFEREE_BONUS: "Referee Bonus",
-    MANUAL_ADJUSTMENT: "Manual Adjustment",
-  };
-
-  function formatDate(dateStr: string): string {
-    return new Date(dateStr).toLocaleDateString("en-IN", {
-      day: "numeric",
-      month: "short",
-      year: "numeric",
-    });
-  }
-
-  function formatAmount(paise: number) {
-    return new Intl.NumberFormat("en-IN", {
-      style: "currency",
-      currency: "INR",
-    }).format(paise / 100);
-  }
+  const creditColumns: ResponsiveColumn<CreditData["history"][number]>[] = [
+    {
+      key: "source",
+      header: "Source",
+      primary: true,
+      cell: (credit) =>
+        creditSourceLabels[credit.source] ?? credit.source.replace(/_/g, " "),
+    },
+    {
+      key: "amount",
+      header: "Amount",
+      cell: (credit) => formatAmount(credit.amount),
+    },
+    {
+      key: "remaining",
+      header: "Remaining",
+      cell: (credit) => formatAmount(credit.remainingAmount),
+    },
+    {
+      key: "expires",
+      header: "Expires",
+      cell: (credit) => (
+        <span className="text-zinc-500">
+          {credit.expiresAt ? formatDate(credit.expiresAt) : "Never"}
+        </span>
+      ),
+    },
+  ];
 
   return (
     <>
@@ -179,139 +258,139 @@ export default function ConsultantReferralsPage({
         subtitle="Invite others and earn credits when they make their first booking"
       />
       <DashboardContent>
-        <DashboardGrid columns={4}>
-          <StatCard
-            title="Total Referred"
-            value={totalReferred}
-            icon={Users}
-          />
-          <StatCard
-            title="Qualified"
-            value={qualified}
-            icon={Gift}
-            variant="success"
-            tooltip={`Friends who signed up and completed their first paid booking within ${QUALIFICATION_WINDOW_DAYS} days`}
-          />
-          <StatCard
-            title="Total Earned"
-            value={formatAmount(code?.totalEarned ?? 0)}
-            icon={IndianRupee}
-          />
-          <StatCard
-            title="Credit Balance"
-            value={formatAmount(credits?.totalAvailable ?? 0)}
-            icon={IndianRupee}
-            variant="info"
-            tooltip="Credits earned from referrals. Applied at checkout."
-          />
-        </DashboardGrid>
+        {statsLoading && !code && !referralsData ? (
+          <DashboardGrid columns={4}>
+            {[1, 2, 3, 4].map((i) => (
+              <StatCardSkeleton key={i} />
+            ))}
+          </DashboardGrid>
+        ) : (
+          <DashboardGrid columns={4}>
+            <StatCard
+              title="Total Referred"
+              value={statsError ? "—" : totalReferred}
+              icon={Users}
+            />
+            <StatCard
+              title="Qualified"
+              value={statsError ? "—" : qualified}
+              icon={Gift}
+              variant="success"
+              tooltip={`Friends who signed up and completed their first paid booking within ${QUALIFICATION_WINDOW_DAYS} days`}
+            />
+            <StatCard
+              title="Total Earned"
+              value={formatAmount(code?.totalEarned ?? 0)}
+              icon={IndianRupee}
+            />
+            <StatCard
+              title="Credit Balance"
+              value={formatAmount(credits?.totalAvailable ?? 0)}
+              icon={IndianRupee}
+              variant="info"
+              tooltip="Credits earned from referrals. Applied at checkout."
+            />
+          </DashboardGrid>
+        )}
 
         {/* Referral Link */}
         <div className="mt-6 bg-white rounded-xl border border-zinc-200 p-6">
           <h3 className="text-sm font-medium text-zinc-900 mb-3">
             Your Referral Link
           </h3>
-          {code && (
+          {codeError ? (
+            <EmptyState
+              icon={AlertTriangle}
+              title="Couldn't generate your referral link"
+              description="Something went wrong while creating your link. Retry, or come back later."
+              action={
+                <Button variant="outline" onClick={() => refetchCode()}>
+                  Retry
+                </Button>
+              }
+            />
+          ) : (
             <>
-              <div className="mb-3 rounded-lg bg-emerald-50 border border-emerald-200 px-4 py-3 text-sm text-emerald-800">
-                Earn {formatAmount(code.referrerReward)} for each friend who
-                books. Your friend gets {formatAmount(code.refereeReward)} off
-                their first booking!
+              {code && (
+                <>
+                  <div className="mb-3 rounded-lg bg-emerald-50 border border-emerald-200 px-4 py-3 text-sm text-emerald-800">
+                    Earn {formatAmount(code.referrerReward)} for each friend who
+                    books. Your friend gets {formatAmount(code.refereeReward)}{" "}
+                    off their first booking!
+                  </div>
+                  <div className="flex items-center gap-2 mb-3">
+                    <span className="text-sm text-zinc-500">Your code:</span>
+                    <span className="font-mono font-semibold text-zinc-900 bg-zinc-100 px-3 py-1 rounded-md">
+                      {code.customCode || code.code}
+                    </span>
+                  </div>
+                </>
+              )}
+              <div className="flex gap-2">
+                <div className="flex-1 flex items-center rounded-md border border-zinc-200 bg-zinc-50 px-3 py-2 font-mono text-sm text-zinc-800 select-all truncate">
+                  {referralLink || (
+                    <span className="text-zinc-400 italic font-sans">
+                      {codeLoading ? "Generating link..." : "No link available"}
+                    </span>
+                  )}
+                </div>
+                <Button
+                  variant="outline"
+                  size="icon"
+                  onClick={handleCopy}
+                  disabled={!referralLink}
+                >
+                  {copied ? (
+                    <Check className="h-4 w-4" />
+                  ) : (
+                    <Copy className="h-4 w-4" />
+                  )}
+                </Button>
+                <Button
+                  variant="outline"
+                  size="icon"
+                  asChild
+                  disabled={!referralLink}
+                >
+                  <a
+                    href={whatsappUrl}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    aria-label="Share on WhatsApp"
+                  >
+                    <WhatsAppIcon className="h-4 w-4" />
+                  </a>
+                </Button>
+                <Button
+                  variant="outline"
+                  size="icon"
+                  asChild
+                  disabled={!referralLink}
+                >
+                  <a
+                    href={emailUrl}
+                    target="_blank"
+                    rel="noopener noreferrer"
+                    aria-label="Share via email"
+                  >
+                    <Mail className="h-4 w-4" />
+                  </a>
+                </Button>
               </div>
-              <div className="flex items-center gap-2 mb-3">
-                <span className="text-sm text-zinc-500">Your code:</span>
-                <span className="font-mono font-semibold text-zinc-900 bg-zinc-100 px-3 py-1 rounded-md">
-                  {code.customCode || code.code}
-                </span>
+              <div className="flex gap-2 mt-3">
+                <Input
+                  placeholder="Set custom code (e.g. MYNAME)"
+                  value={customCode}
+                  onChange={(e) => setCustomCode(e.target.value)}
+                  className="flex-1"
+                />
+                <Button variant="outline" onClick={handleCustomize}>
+                  Set
+                </Button>
               </div>
             </>
           )}
-          <div className="flex gap-2">
-            <div className="flex-1 flex items-center rounded-md border border-zinc-200 bg-zinc-50 px-3 py-2 font-mono text-sm text-zinc-800 select-all truncate">
-              {referralLink || (
-                <span className="text-zinc-400 italic font-sans">
-                  Generating link...
-                </span>
-              )}
-            </div>
-            <Button variant="outline" size="icon" onClick={handleCopy}>
-              {copied ? (
-                <Check className="h-4 w-4" />
-              ) : (
-                <Copy className="h-4 w-4" />
-              )}
-            </Button>
-            <Button
-              variant="outline"
-              size="icon"
-              asChild
-              disabled={!referralLink}
-            >
-              <a
-                href={whatsappUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                aria-label="Share on WhatsApp"
-              >
-                <WhatsAppIcon className="h-4 w-4" />
-              </a>
-            </Button>
-            <Button
-              variant="outline"
-              size="icon"
-              asChild
-              disabled={!referralLink}
-            >
-              <a
-                href={emailUrl}
-                target="_blank"
-                rel="noopener noreferrer"
-                aria-label="Share via email"
-              >
-                <Mail className="h-4 w-4" />
-              </a>
-            </Button>
-          </div>
-          <div className="flex gap-2 mt-3">
-            <Input
-              placeholder="Set custom code (e.g. MYNAME)"
-              value={customCode}
-              onChange={(e) => setCustomCode(e.target.value)}
-              className="flex-1"
-            />
-            <Button variant="outline" onClick={handleCustomize}>
-              Set
-            </Button>
-          </div>
         </div>
-
-        {/* TODO: Program Details - uncomment when ready to show
-        <div className="mt-6 bg-white rounded-xl border border-zinc-200 p-6">
-          <h3 className="text-sm font-medium text-zinc-900 mb-3">
-            Program Details
-          </h3>
-          <div className="grid grid-cols-3 gap-4 text-sm">
-            <div className="rounded-lg bg-zinc-50 px-4 py-3">
-              <p className="text-zinc-500">Qualification Window</p>
-              <p className="mt-1 font-medium text-zinc-900">
-                {QUALIFICATION_WINDOW_DAYS} days
-              </p>
-            </div>
-            <div className="rounded-lg bg-zinc-50 px-4 py-3">
-              <p className="text-zinc-500">Credit Expiry</p>
-              <p className="mt-1 font-medium text-zinc-900">
-                {CREDIT_EXPIRY_DAYS} days
-              </p>
-            </div>
-            <div className="rounded-lg bg-zinc-50 px-4 py-3">
-              <p className="text-zinc-500">Max Referrals</p>
-              <p className="mt-1 font-medium text-zinc-900">
-                {code?.maxReferrals ?? 25}
-              </p>
-            </div>
-          </div>
-        </div>
-        */}
 
         {/* Referrals List */}
         <div className="mt-6 bg-white rounded-xl border border-zinc-200 overflow-hidden">
@@ -320,93 +399,74 @@ export default function ConsultantReferralsPage({
               Your Referrals
             </h3>
           </div>
-          {referrals.length === 0 ? (
-            <div className="px-6 py-12 text-center text-zinc-500 text-sm">
-              No referrals yet. Share your link to get started!
+          {referralsLoading && !referralsData ? (
+            <div className="p-4">
+              <DataCardSkeleton />
             </div>
+          ) : referralsError ? (
+            <EmptyState
+              icon={AlertTriangle}
+              title="Couldn't load your referrals"
+              description="Retry, or come back later."
+              action={
+                <Button variant="outline" onClick={() => refetchReferrals()}>
+                  Retry
+                </Button>
+              }
+            />
           ) : (
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="border-b border-zinc-100 text-zinc-500">
-                  <th className="px-6 py-3 text-left font-medium">User</th>
-                  <th className="px-6 py-3 text-left font-medium">Status</th>
-                  <th className="px-6 py-3 text-left font-medium">Signed Up</th>
-                  <th className="px-6 py-3 text-right font-medium">Reward</th>
-                </tr>
-              </thead>
-              <tbody>
-                {referrals.map((ref) => (
-                  <tr
-                    key={ref.id}
-                    className="border-b border-zinc-50 last:border-0"
-                  >
-                    <td className="px-6 py-3">{ref.referredUser.name}</td>
-                    <td className="px-6 py-3">
-                      <span
-                        className={`inline-flex items-center rounded-full px-2 py-1 text-xs font-medium ${
-                          referralStatusConfig[ref.status]?.className ??
-                          "bg-zinc-100 text-zinc-600"
-                        }`}
-                      >
-                        {referralStatusConfig[ref.status]?.label ?? ref.status}
-                      </span>
-                    </td>
-                    <td className="px-6 py-3 text-zinc-500">
-                      {formatDate(ref.signedUpAt)}
-                    </td>
-                    <td className="px-6 py-3 text-right">
-                      {ref.status === "REWARDED"
-                        ? formatAmount(ref.referrerRewardAmount)
-                        : "-"}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
+            <ResponsiveTable
+              columns={referralColumns}
+              rows={referrals}
+              getRowId={(ref) => ref.id}
+              className="[&>ul]:p-3"
+              empty={
+                <EmptyState
+                  icon={Users}
+                  title="No referrals yet"
+                  description="Share your link to get started!"
+                />
+              }
+            />
           )}
         </div>
 
         {/* Credit History */}
-        {credits?.history && credits.history.length > 0 && (
+        {creditsError ? (
           <div className="mt-6 bg-white rounded-xl border border-zinc-200 overflow-hidden">
             <div className="px-6 py-4 border-b border-zinc-200">
               <h3 className="text-sm font-medium text-zinc-900">
                 Credit History
               </h3>
             </div>
-            <table className="w-full text-sm">
-              <thead>
-                <tr className="border-b border-zinc-100 text-zinc-500">
-                  <th className="px-6 py-3 text-left font-medium">Source</th>
-                  <th className="px-6 py-3 text-left font-medium">Amount</th>
-                  <th className="px-6 py-3 text-left font-medium">Remaining</th>
-                  <th className="px-6 py-3 text-left font-medium">Expires</th>
-                </tr>
-              </thead>
-              <tbody>
-                {credits.history.map((credit) => (
-                  <tr
-                    key={credit.id}
-                    className="border-b border-zinc-50 last:border-0"
-                  >
-                    <td className="px-6 py-3">
-                      {creditSourceLabels[credit.source] ??
-                        credit.source.replace(/_/g, " ")}
-                    </td>
-                    <td className="px-6 py-3">{formatAmount(credit.amount)}</td>
-                    <td className="px-6 py-3">
-                      {formatAmount(credit.remainingAmount)}
-                    </td>
-                    <td className="px-6 py-3 text-zinc-500">
-                      {credit.expiresAt
-                        ? formatDate(credit.expiresAt)
-                        : "Never"}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
+            <EmptyState
+              icon={AlertTriangle}
+              title="Couldn't load your credits"
+              description="Retry, or come back later."
+              action={
+                <Button variant="outline" onClick={() => refetchCredits()}>
+                  Retry
+                </Button>
+              }
+            />
           </div>
+        ) : (
+          credits?.history &&
+          credits.history.length > 0 && (
+            <div className="mt-6 bg-white rounded-xl border border-zinc-200 overflow-hidden">
+              <div className="px-6 py-4 border-b border-zinc-200">
+                <h3 className="text-sm font-medium text-zinc-900">
+                  Credit History
+                </h3>
+              </div>
+              <ResponsiveTable
+                columns={creditColumns}
+                rows={credits.history}
+                getRowId={(credit) => credit.id}
+                className="[&>ul]:p-3"
+              />
+            </div>
+          )
         )}
       </DashboardContent>
     </>
